LayoutTests/platform/gtk/storage/domstorage/sessionstorage/blocked-file-access-expected.txt

CONSOLE MESSAGE: line 10: PASS: window.sessionStorage was accessible

Test that we are permitted access to sessionStorage from a file URL if universal access is turned off.
